backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.7
langchain-chroma==0.1.4
chromadb==0.5.20
pypdf==5.1.0
python-multipart==0.0.12
openai==1.55.3
tiktoken==0.8.0
boto3==1.35.71
